backend/requirements.txt
asgiref==3.8.1
Django==3.1.12
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
pymongo==3.11.4
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.2.4
typing_extensions==4.12.2
